Addition of AssemblyExtensions
Add Assembly Extensions Class:

- Includes all assembly methods that do not not need firebase configuration.
