+++ src/compas_xr/assemblies/__init__.py
@@ -1,3 +1,4 @@
 from .assemblies import AssemblyAssistant
+from .assembly_extensions import AssemblyExtensions
 
-__all__ = ["AssemblyAssistant"]
+__all__ = ["AssemblyAssistant", "AssemblyExtensions"]
