+++ src/compas_xr/storage/storage_interface.py
@@ -17,3 +17,6 @@ class StorageInterface(object):
 
     def upload_objs(self, folder_local, cloud_folder_name):
         raise NotImplementedError("Implemented on child classes")
+
+    # def upload_graph(self, graph):
+    #     upload_data(graph.attributes[]
